velkozz_logger/requirements.txt
Flask
pandas
networkx
flask-restful
plotly
Flask-WTF
WTForms
Flask-SQLAlchemy
python-dotenv